Fix some bugs

+++ src/Chee/Module/CheeModule.php
@@ -595,13 +595,16 @@ class CheeModule
             $this->app['events']->fire('modules.uninstall.'.$name, null);
             $this->app['events']->fire('modules.delete.'.$name, null);
             $module->delete();
-            if (!$this->files->deleteDirectory($this->getAssetDirectory($name)))
+            if ($this->exists($this->getAssetDirectory($name)))
             {
-                $this->errors['delete']['forbidden'] = $this->getAssetDirectory($name);
+                if (!$this->files->deleteDirectory($this->getAssetDirectory($name)))
+                {
+                    $this->errors['delete']['forbidden']['assest'] = $this->getAssetDirectory($name);
+                }
             }
             if (!$this->files->deleteDirectory($this->getModuleDirectory($name)))
             {
-                $this->errors['delete']['forbidden'] = $this->getModuleDirectory($name);
+                $this->errors['delete']['forbidden']['module'] = $this->getModuleDirectory($name);
             }
             return true;
         }
